[FEAT] Fix PeepMiniCalendar Design and remove unnecessary parameters
