+++ lib/app/views/test.dart
@@ -11,7 +11,7 @@ import 'package:peep_todo_flutter/app/theme/icons.dart';
 import 'package:peep_todo_flutter/app/theme/palette.dart';
 import 'package:peep_todo_flutter/app/views/common/peep_dropdown_menu.dart';
 import 'package:peep_todo_flutter/app/views/main/widget/peep_scheduled_todo_app_bar.dart';
-import 'package:peep_todo_flutter/app/views/main/widget/peep_week_calendar_bar.dart';
+import 'package:peep_todo_flutter/app/views/todo/widget/peep_mini_calendar.dart';
 import 'package:table_calendar/table_calendar.dart';
 import 'common/todo/peep_todo_list.dart';
 
@@ -34,41 +34,6 @@ class Test extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return PeepWeekCalendarBar(
-      dropdownMenuItems: [
-        DropdownMenuItemData(
-            'popup_action_1',
-            PeepIcon(Iconsax.addcircle,
-                size: AppValues.smallIconSize, color: Palette.peepBlack),
-            '카테고리 추가'),
-        DropdownMenuItemData(
-            'popup_action_2',
-            PeepIcon(Iconsax.categorybox,
-                size: AppValues.smallIconSize, color: Palette.peepBlack),
-            '카테고리 관리'),
-        DropdownMenuItemData(
-            'popup_action_3',
-            PeepIcon(Iconsax.reminder,
-                size: AppValues.smallIconSize, color: Palette.peepBlack),
-            '리마인더 관리'),
-        DropdownMenuItemData(
-            'popup_action_4',
-            PeepIcon(Iconsax.routine,
-                size: AppValues.smallIconSize, color: Palette.peepBlack),
-            '루틴 추가'),
-      ],
-      onMenuItemSelected: (popupNum) {
-        if (popupNum == 'popup_action_1') {
-          onTapFunc() => print('1');
-        } else if (popupNum == 'popup_action_2') {
-          onTapSecondFunc() => print('2');
-        } else if (popupNum == 'popup_action_3') {
-          o3() => print('3');
-        } else {
-          o4() => print('4');
-        }
-      },
-      onTapClock: () => print('123'),
-    );
+    return PeepMiniCalendar();
   }
 }
